Move classes to files of their own
